% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_basemap}
\alias{get_basemap}
\title{Generate basemap for ZCTA boundaries}
\usage{
get_basemap(zcta_list, year_range)
}
\arguments{
\item{zcta_list}{Vector of ZCTAs (or ZCTA prefixes)}

\item{year_range}{Range of years across which data should be aggregated}
}
\value{
A mapping of boundaries for requested ZCTAs
}
\description{
Takes a list of US ZCTA codes and a year, and returns a plot of the requested
ZCTA boundaries
}
\note{
ZCTAs/prefixes can be passed in as characters or integers

ZCTAs/prefixes can be 1-5 characters
}
\keyword{internal}
