% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_data}
\alias{get_data}
\title{Generates a dataframe containing all exposures for a given set of ZCTA-years}
\usage{
get_data(zcta_list, year_range)
}
\arguments{
\item{zcta_list}{Vector of ZCTAs (or ZCTA prefixes)}

\item{year_range}{Range of years across which data should be aggregated}
}
\value{
A dataframe with exposure data for selected ZCTAs
}
\description{
Takes a list of US ZCTA codes, a year, and a feature, and returns a dataframe
containing all exposure data for the requested ZCTA boundaries
}
\note{
ZCTAs/prefixes can be passed in as characters or integers

ZCTAs/prefixes can be 1-5 characters
}
