% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_geometry}
\alias{get_geometry}
\title{Retrieve ZCTA geometries from Tigris}
\usage{
get_geometry(zcta_list, year_range)
}
\arguments{
\item{zcta_list}{Vector of ZCTAs (or ZCTA prefixes)}

\item{year_range}{Range of years across which data should be aggregated}
}
\value{
An sf object containing ZCTA boundary geometries
}
\description{
Takes a list of US ZCTA codes and a year, and returns a simple features
geometry object for those ZCTA boundaries in the given year
}
\note{
ZCTAs/prefixes can be passed in as characters or integers

ZCTAs/prefixes can be 1-5 characters
}
\keyword{internal}
