% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{add_tracks}
\alias{add_tracks}
\title{Overlay tornado tracks on top of ZCTA boundary maps}
\usage{
add_tracks(zcta_list, year_range, plot)
}
\arguments{
\item{zcta_list}{Vector of ZCTAs (or ZCTA prefixes)}

\item{year_range}{Range of years across which data should be aggregated}

\item{plot}{An sf plot object}
}
\value{
A map with tornado tracks overlaid
}
\description{
Takes a map object (either choropleth filled or blank ZCTA boundaries) and
overlays tornado tracks
}
\note{
ZCTAs/prefixes can be passed in as characters or integers

ZCTAs/prefixes can be 1-5 characters

Can be a choropleth or map of ZCTA boundaries created by ```map_exposure```
}
