% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{generate_feature}
\alias{generate_feature}
\title{Aggregates a feature of interest across ZCTAs for given timeframe}
\usage{
generate_feature(exposed_zctas, feature = c("tornado_id", "mag", "fat", "inj"))
}
\arguments{
\item{exposed_zctas}{Dataframe of tornado-level exposures}

\item{feature}{The feature of interest (tornado_id for count, mag for magnitude,
fat for fatalities, inj for injuries)}
}
\value{
Dataframe of feature aggregated at ZCTA level
}
\description{
Takes a dataframe of exposures, returns a dataframe where each row is a ZCTA
and the sum or average (if magnitude) of the feature
}
\keyword{internal}
