% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{map_exposure}
\alias{map_exposure}
\title{Create a choropleth map for variable of interest across selected ZCTAs}
\usage{
map_exposure(zcta_list, year_range, feature)
}
\arguments{
\item{zcta_list}{Vector of ZCTAs (or ZCTA prefixes)}

\item{year_range}{Range of years across which data should be aggregated}

\item{feature}{Name of feature to be visualized (can be tornado_id, mag,
fatality, injury)}
}
\value{
A map of the distribution of feature of interest across selected ZCTAs
}
\description{
Takes a list of US ZCTA codes, a year, and a feature, and returns a plot of
the requested ZCTA boundaries with a choropleth fill to represent the spatial
distribution of the feature
}
\note{
ZCTAs/prefixes can be passed in as characters or integers

ZCTAs/prefixes can be 1-5 characters

Feature name should align with column name in dataset, must be string
}
